feat: add tests
